+++ panda.config.ts
@@ -1,8 +1,8 @@
 import { keyframes } from '@/styles/keyframes';
-import { colorVariables } from '@/ui/styled/CssBaseline/Settings/Color';
-import { easingVariables } from '@/ui/styled/CssBaseline/Settings/Easing';
-import { fontVariables } from '@/ui/styled/CssBaseline/Settings/Font';
-import { spaceVariables } from '@/ui/styled/CssBaseline/Settings/Space';
+import { colorVariables } from '@/ui/styled/variables/Color';
+import { easingVariables } from '@/ui/styled/variables/Easing';
+import { fontVariables } from '@/ui/styled/variables/Font';
+import { spaceVariables } from '@/ui/styled/variables/Space';
 import { defineConfig } from '@pandacss/dev';
 
 export default defineConfig({
