+++ src/components/UI/Layout/Cluster.tsx
@@ -1,8 +1,8 @@
 import type { AriaRole, ReactNode } from 'react';
 import { memo, useMemo } from 'react';
 
-import type { SpaceGap } from '@/ui/styled/CssBaseline/Settings/Space';
 import { css, styled } from '@/ui/styled/dynamic';
+import type { SpaceGap } from '@/ui/styled/variables/Space';
 
 type Props = {
   as?: keyof JSX.IntrinsicElements;
