+++ src/ui/styled/variables/Easing.ts
